remove logs associated with warning events
Generating logs for Warning events ends up generating 2 logs for each
event; 1 for the event itself and another for the reconciler error that
is always related to a Warning event.

To reduce the number of redundant logs we are changing the event logger
interface to set the warning event log to debug so that it only comes up
if the log level is set appropriately.  For normal events, since they
are never associated to a reconciler error they need to produce a log so
that we can see them in the log stream rather than having to following
the kubernetes event stream which is less convenient because of how it
deals with repeating events on the same resource.

Signed-off-by: Allain Legacy <allain.legacy@windriver.com>