fix issues related to referencing loop variables
Taking a pointer to a field of a loop variable is unsafe because the
value will change on the next loop iteration but the entity pointing to
it will assume that the previous value still remained.

This caused interfaces to have incorrect MTU values as their values
would changed based on other MTU values of other interfaces processed in
the loop.  This was not seen previously as all tested lab configurations
had identical MTU values for all interfaces.

This commit changes the specific code related to the bug but also
changes the object constructors to use pass-by-value so that any
pointers taken for the purpose of creating API objects are safe to use
beyond the scopy of the inventory cache objects.

Signed-off-by: Allain Legacy <allain.legacy@windriver.com>