remove private data from collected defaults
This commit ensures that system private data (certificate private keys
and bmc password) information is excluded from the "defaults" that are
collected and stored on the resource status.  The private data is not
actually stored as part of the defaults, but a placeholder to it is
saved and since nothing is going to populate the related Secrets this
causes reconciliation errors when trying to resolve the difference
between the stored defaults and the running configuration.

We will never be able to rollback the system certificates or bmc
information back to the original values so there is no point in storing
them in the defaults.

Signed-off-by: Allain Legacy <allain.legacy@windriver.com>