reorder copy instructions in dockerfile
The order of the copy operations in the Dockerfile have been changed to
ensure that the least likely files to be modified are copied first.
This guarantees that incremental builds are a little faster since
unchanged layers are pulled from the cache.  The "vendor" directory has
the biggest impact since it rarely ever changes and is the largest
size directory to be copied.

Signed-off-by: Allain Legacy <allain.legacy@windriver.com>