add trusted ca system filter to deploy tool
This adds a new trusted CA filter to the deploy build tool.  The purpose
is to filter out the "ssl_ca" and "openstack_ca" certificates from the
generated deployment-config.yaml file.  This is because they are more
likely to have been added to the system at bootstrap time and so aren't
necessary in the deployment configuration.  They can be added to the
deployment config but since the certificate hash will be the same it
won't result in any changes by the reconciler so it is just extra steps
for the end user to have to add the x.509 data to the config.  Therefore
we are removing these as part of the --minimal-config filters on the
build tool for now.

Signed-off-by: Allain Legacy <allain.legacy@windriver.com>