add a new state change monitor for reconciled hosts
It is possible for a host to change state autonomously (or by another
API user) even after the deployment manager has finished reconciling the
host.  This causes the local cached copy of the host state to be out of
date and may cause confusion for end users.

This commit introduces a new host state monitor that is launched after
the host has been reconciled so that state changes are tracked.  This is
not critical so the frequency is set to 2 minutes to reduce API usage
once host reconciliation completes.

Signed-off-by: Allain Legacy <allain.legacy@windriver.com>