add support for host fs configuration
This commit splits the existing controller filesystem function so that
replicated controller filesystem remain as configuration items at the
system level and non-replicated host filesystems are moved to a separate
host storage filesystem section.  This is to align with the recent
changes to the system API that implemented the equivalent change.

Signed-off-by: Allain Legacy <allain.legacy@windriver.com>