wait for host filesystems before collecting defaults
The current method of waiting for inventory collection is to poll the
disk list on a host and when it is non-empty we proceed with collecting
default values.  With the introduction of the host filesystem feature
the inventory collection should not be considered complete until the
host filesystem list is non-empty.

Signed-off-by: Allain Legacy <allain.legacy@windriver.com>