merge filesystem lists instead of replacing
If a user only specifies a subset of the host filesystems because only
that subset needs to be altered from the system defaults then we should
replace those values in the default host filesystem list rather than
replace the entire list with the user overrides.  It is important to do
this because when comparing the current config to the desired config we
need to ensure that existing/unchange filesystems do not trigger a
comparison failure.

By adding a IsKeyEqual method to the host and controller filesystem data
structures we ensure that the merge transformer will merge them properly
rather than replace the defaults with the user overrides.

Signed-off-by: Allain Legacy <allain.legacy@windriver.com>