fix crash when merging from nil slice pointer
The previous fix to address merging both slice and slice pointers did
not validate whether the pointers where nil before dereferencing them.
This lead to a segfault of the deployment manager.

This commit addresses the issue by first checking that the pointer is
non-nil and dealing with it properly if it is.  Unit tests were added to
ensure that no regression slip in for this area in the future.

Signed-off-by: Allain Legacy <allain.legacy@windriver.com>