support multiple auth url values
Some systems netboot with an IP address that is different than the
eventual OAM floating IP address that will be configured.  In those
scenarios, the deployment manager loses connectivity to the system when
the IP address switches to the OAM floating IP so in order to make the
deployment manager aware of both addresses we are supporting a comma
separated list of auth URL values in OS_AUTH_URL.  It is important that
the OAM floating IP be placed in the list ahead of the temporary IP
address to avoid 30 seconds delays each time the deployment manager
needs to reconnect to the system subsequent to the system being
reconfigured with the OAM IP address.

Signed-off-by: Allain Legacy <allain.legacy@windriver.com>