set default repository path to be local
Since we do not yet have a publically available docker registry set the
default repository path to be local.  Images can be built an loaded
manually onto the system being deployed, or an end user can store the
image into their own private registry and then override the helm chart
to point to that private registry.

Signed-off-by: Allain Legacy <allain.legacy@windriver.com>