fail interface update if missing network resource
If a network or datanetwork is missing during interface reconciliation
then an error should be raised and reconciliation should stop.
Currently while adding a interface-network or interface-datanetwork
association a missing resource is simply ignored.  This results in the
node being unlocked anyway and then the system forever thinks that the
host is out-of-sync because of a missing network association or
datanetwork association.

Signed-off-by: Allain Legacy <allain.legacy@windriver.com>