rework network association move operations
When a network is moved from one interface to another it first needs to
be removed from the old interface otherwise an API error may occur as
networks can only be assigned to a single interface at a time.  The
current implementation attempted to reconcile all interface-network
changes in a single pass and would fail whenever a network was moved.

The new implementation executes this task in 2 separate phases.  First,
the network associations are removed from interfaces that no longer
have them configured and then during a second pass network associations
are added to interfaces that have them configured.

This commit also addresses the exact same issue with datanetwork
associations.

Signed-off-by: Allain Legacy <allain.legacy@windriver.com>