refactor merge transformers and add unit tests
Refactoring the merge transformers to break the lambda functions out of
the main tranformer and into custom functions.

This commit also adds unit tests to ensure that going forward there will
be no accidental regressions on the merge function.  This is a critical
piece of code and tends to be tricky to get right in some cases.  The
impact of getting it wrong is that unpredictable (and difficult to
debug) behaviour may occur at runtime.

Signed-off-by: Allain Legacy <allain.legacy@windriver.com>