add support for logging all rest api requests
If the OS_DEBUG attribute is set to a boolean value then all REST API
requests and responses between the Deployment Manager and the StarlingX
system API will be logged to stdout.  The value is parsed using standard
Go libraries therefore the value used must be understood as a boolean by
strconv.ParseBool which at the time of this commit is "1", "t", "T",
"true", "TRUE", and "True".

Signed-off-by: Allain Legacy <allain.legacy@windriver.com>