prevent host provisioning after initial synchronization
The intent of the design is to bring a host to desired configuration
state and then to stop manipulating the host.  There was already code
in place to prevent any attribute changes on a host that has already
reached the desired configurations state, but no similar check was put
in place to stop a host from being re-provisioned if it was manually
deleted by the end user at the system API.  This commit prevents
a host from being provisioned again if it was deleted by the end user
after it had already been configured completely.

Signed-off-by: Allain Legacy <allain.legacy@windriver.com>