re-enable memory configuration reconciler
The memory reconciler was disabled because the memory values reported by
the system API changed dynamically at runtime thus causing the
deployment manager to think that the system was out of sync with the
desired configuration.  The system API has been fixed so that the values
reported are constant and will not change unless a configuration change
is applied therefore it is safe to re-enable the memory reconciler.

Signed-off-by: Allain Legacy <allain.legacy@windriver.com>