access region name from secret data attribute
The StringData attribute of a Secret is a write-only field.  Any values
written to StringData[x] should be read from Data[x].  Since the region
name attribute was being accessed from the wrong attribute name the
value was always returned as an empty string and so resulted in an
error when accessing endpoints with individual values for each
distinct region.

Signed-off-by: Allain Legacy <allain.legacy@windriver.com>