move bm validation from webhook to runtime
Since the board management info can be spread over multiple layered
profiles and/or host overrides it doesn't make sense to validate the
attributes in the webhook since they may not all be present.  That is, a
top level profile my have the "type", but a lower profile may have the
"credentials".  It is not until the composite profile is built that all
related attributes would be present so delay the validation until that
point.

Signed-off-by: Allain Legacy <allain.legacy@windriver.com>