add reconciled attribute to other resources
The Host resource supports a "reconciled" attribute in its Status to
track whether it has reached the InSync status at least once.  The other
resources will require this type of functionality so this is a first
pass at adding the additional attribute.  Actual usuage of the attribute
will come in a separate commit.

Signed-off-by: Allain Legacy <allain.legacy@windriver.com>