add license support to the system resource
This commit adds the ability to specify the contents of a license file
as a Secret and to refer to that Secret from the System resource.  The
license contents are installed at the system API if the system does not
already have a license installed or if the contents of the current and
proposed licenses do not match.

Signed-off-by: Allain Legacy <allain.legacy@windriver.com>