remove unneeded decoding of certificate data
The handling of the root certificates is not working because the Secret
data does not need to be decoded.  The data is already presented in raw
form rather than in base64 encoded form.  This code was inadvertantly
working properly previously because there was a missing return code
check on the Decode() operation.  That operation was failing but the
code path was continuing and using the raw data.  The missing return was
identified by static analysis and from that point on root certificate
installs must have been failing.

Signed-off-by: Allain Legacy <allain.legacy@windriver.com>