publish first official released version
This publishes the first DM chart and image as v1.0.0.  A support branch
will be created for issues related to this version, and the master
branch will be immediately promoted to v1.1.0 to allow for development
of new features on the master branch while allowing v1.0.0 to continue
in support mode on its own branch.

Signed-off-by: Allain Legacy <allain.legacy@windriver.com>