update osd handling to address missing journals
The OSD handling assumed that each OSD would be associated with a
cluster, but Journal OSDs are not therefore they should not be
filtered out based on this criteria.

The OSD handling of the "journal_size_mib" attribute was incorrectly
using directly as GiB rather than converting from MiB to GiB therefore
the deploy tool was generating YAML that was invalid when re-applied to
the system.

Signed-off-by: Allain Legacy <allain.legacy@windriver.com>