package helm chart only when necessary
The Helm package commands makes no attempt at determining whether
the chart stored in the output location is equivalent to the one that
is about to be packaged.  This leads to accidentally storing updated
Helm charts into the repo.

This commit updates the makefile so that it only runs the Helm package
command whenever one of the dependent files is updated.  If the package
must be re-built regardless of the contents then the HELM_FORCE=1
variable can be used to override the makefile behaviour.

make helm-package
or
HELM_FORCE=1 helm-package

Signed-off-by: Allain Legacy <allain.legacy@windriver.com>