add build container to allow building from pipelines
Building various artifacts for the deployment manager and running unit
tests requires specific packages being installed on the host machine.
To enable building these artifacts from an automated pipeline (e.g.,
jenkins) we are providing a "builder" container which can be run to
execute the unit tests and other build steps within a container.

The intent is that the "builder" container be run with the local copy
of the repo mounted as a volume within the container so that the source
code is available to the container and so that build output can be
stored within the repo just like if the build was run locally (outside
of a build container).

The builder container also should be executed with the the docker
socket (/var/run/docker.sock) mounted as a volume so that the docker
command running inside the container can build and publish images to the
host's docker environment (i.e., docker-in-docker, or dind).

Signed-off-by: Allain Legacy <allain.legacy@windriver.com>