add stable state host monitor
Monitoring of the host state for transient states was accomplished
with the idle state monitor.  That monitor only considered the "Task"
attribute of hosts to determine if the host was transitioning to some
other state.  It did not consider state transitions reflected by a
mismatch between the admin state and the operational state.  That is, if
a host is unlocked/disabled then it is not stable in the sense that is
not a final state.  It is either moving towards the locked/disabled
state or to the unlocked/enabled state.

To avoid making configuration decisions while a transition is in
progress a new "stable" host monitor is being used that will consider
admin/operational states as well as any currently in progress tasks.

Signed-off-by: Allain Legacy <allain.legacy@windriver.com>