update default auth url ip address in examples
StarlingX has recently updated its default IP allocation scheme for the
management network so that the floating IP address is now using the .1
address rather than the .2 address.  We are updating our examples so
that anyone using config files based on our examples will not be using
an invalid configuration out of the box.

Signed-off-by: Allain Legacy <allain.legacy@windriver.com>