Fix bm.Credentials dereference error in deploy tool
Recent updates made the credentials and address fields of
BoardManagement optional, if bmType is "none". The deploy tool had an
assumption that if BoardManagement is set, then Credentials was as
well. In the case of bmType=none, this led to dereferencing a null
pointer.

This commit adds a nil check to test the Credentials field before the
dereference.

Signed-off-by: Don Penney <don.penney@windriver.com>