Update hugepage count parsing
The memory reconciliation is checking the pending hugepage count data,
to account for memory configuration that has been issued but not yet
completed. The parsing of current config, however, was not doing the
same. As a result, after initial unlock, the hugepage count from the
system could be coming back as 0 with a change pending, but this will
just look as though the system has a current config of 0 hugepages.
This causes the host to remain out-of-sync, as it doesn't match
expected ocnfiguration.

This update applies the same logic in the memory parser as the
reconciler, setting the hugepage count based on pending data first, if
available, and to the actual hugepage count if no update is pending.

Signed-off-by: Don Penney <don.penney@windriver.com>