Add support for ethernet type interface on ethernet interface
In order to support configuring CaaS and Data networks to single VF
capable NIC, we add support for creating ethernet type interface on
top of ethernet interface which is actually class of pci-sriov.
https://storyboard.openstack.org/#!/story/2008470

This way, on top of a same pci-sriov ethernet interface, we can:
- create an untagged ethernet interface for pxeboot platform network.
- create vlan tagged interfaces for oam, internal-mgmt and cluster-host
- create vf type sub-interfaces for data networks

For stable/v2.0.5 branch, charts version has been updated to 2.0.5-2.

Signed-off-by: Litao Gao <litao.gao@windriver.com>
Signed-off-by: Don Penney <don.penney@windriver.com>

Conflicts:
	docs/charts/index.yaml
	docs/charts/wind-river-cloud-platform-deployment-manager-2.0.6.tgz