Update minimum required helm version to v2.16.10
An update in helm v2.16 fixed an issue generating charts where
symbolically linked directories are used. Earlier versions would end
up traversing the linked directories twice, resulting in a larger
charts tarball with extra unwanted files.

This commit updates the DEVELOPER guide to change the minimum helm
version to v2.16.10, and also adds a check to the Makefile to ensure
developers using an older helm version must update it.

In addition, the Dockerfile.builder is updated to use this helm
version.

Signed-off-by: Don Penney <don.penney@windriver.com>