Add storage_backend network field
Add support for storage backend network selection
Add the network field to storage_backend endpoint.
The field is only used by ceph at the moment.

Tested creating a docker image.
The value specified in yaml is passed to the REST call.
If the key is missing from the yaml, then no network field is passed
in the REST call.

Change needed for:
https://storyboard.openstack.org/#!/story/2008843
https://review.opendev.org/c/starlingx/config/+/787426

Signed-off-by: Dan Voiculeasa <dan.voiculeasa@windriver.com>