Fix DM reinstall error in helmv2
When DM exists in the helmv2 namespace but helmv3 is used, the
command "helm upgrade --install" will fail. The playbook would try
to install DM in the helmv3 namespace which conflicts with the
existing release in the helmv2 namespace. This change checks which
version of helm the DM release is installed in and uses a separate
reinstallation procedure for helmv2.

Signed-off-by: Melissa Wang <melissa.wang@windriver.com>
(cherry picked from commit 682ff31903bcbf27306f45ac6370b62884274cfa)