Adjust the format of PTP parameters
This commit adjusts the format of the PTP parameters as a list of
strings. This change is to align with the development sysinv APIs of
per instance PTP configuration.
The parameter string will be "key=value".
This commit also removes the description as we don't have the description
value assgined in sysinv db and API.

Signed-off-by: Yuxing Jiang <yuxing.jiang@windriver.com>