Add reconcilers to reconciler the PTP per instance config
This commit adds two new reconcilers for operations on the PTP
instance and the PTP interface, and modifies the existing host
reconciler and network reconciler to the association between the PTP
instance and the host, the PTP interface and the interface
respectively.

Some modifications on the CRDs also included in this commit.

Signed-off-by: Yuxing Jiang <yuxing.jiang@windriver.com>