Add reconcilers to reconciler the PTP per instance config
This commit adds two new reconcilers for operations on the PTP
instance and the PTP interface, and modifies the existing host
reconciler and network reconciler to the association between the PTP
instance and the host, the PTP interface and the interface
respectively.

Some modifications on the CRDs also included in this commit.

Passed test:
1. built the image and tarball
2. installed the DM in an AIOSX system
3. deployed the AIOSX with a ptp instance and a ptp interface
configured in the deploymment-configuration.yaml. The new PTP
instance, interface and params are configured by DM in the system.
After unlocking, so alarm raised.

Signed-off-by: Yuxing Jiang <yuxing.jiang@windriver.com>