Upversion to 2.0.8-1 (#95)
* Waiting for host available after reconciled (#86)

Previously, the host administrative state would be shown as "locked"
for several hours after the controller has been locked and unlocked
after the initial conciliation.

This commit starts a monitor to wait for the host return available
before comparing its config after the initial reconciliation finishes.

Passed tests:

Both controllers are configured as "unlocked" in deployment
configuration:
1 Deploy a DX system. Both controllers are unlocked by DM.
2 Manually lock/unlock the standby controller, check its status
back to unlock/available after it rebooted.

The standby controller is configured as "locked" in deployment
configuration:
1 Deploy a DX system. The controller-0 is locked by DM, the
controller-1 remains locked, both controllers are in-sync and
reconciled
2 Manually unlock the controller-1, its status changes to:
INSYNC: false, RECONCILED: true after it rebooted

Signed-off-by: Yuxing Jiang <yuxing.jiang@windriver.com>

* Add ptp instance and interface resource (#87)

This commit adds the customized PTP instance and interface resource,
add modifies the host profile for PTP configuration.

Test:
1 Build the DM image with this change.
2 Install the DM with the new image and tarball in an AIOSX system.
3 Apply the deployment configuration with ptp instance and interface
with parameters. These resource are validated and created
successfully. And the host config is updated accordingly.

Signed-off-by: Yuxing Jiang <yuxing.jiang@windriver.com>

* Update Gophercloud pkgs (#90)

This commit updates the gophercloud pkgs from:
Wind-River/gopherlcoud:starlingx

Signed-off-by: Yuxing Jiang <yuxing.jiang@windriver.com>

* Add ptpInstance and ptpInterface to constructor (#88)

This update adds new ptp fields to the HostInfo struct and the
host profile specs, which represent the current settings of the
host. With this change, the PTP configuration on a running system
will be picked up by DM.

Signed-off-by: Melissa Wang <melissa.wang@windriver.com>

* Add reconcilers to reconcile the PTP per instance configurations (#89)

* Add reconcilers to reconciler the PTP per instance config

This commit adds two new reconcilers for operations on the PTP
instance and the PTP interface, and modifies the existing host
reconciler and network reconciler to the association between the PTP
instance and the host, the PTP interface and the interface
respectively.

Some modifications on the CRDs also included in this commit.

Passed test:
1. built the image and tarball
2. installed the DM in an AIOSX system
3. deployed the AIOSX with a ptp instance and a ptp interface
configured in the deploymment-configuration.yaml. The new PTP
instance, interface and params are configured by DM in the system.
After unlocking, so alarm raised.

Signed-off-by: Yuxing Jiang <yuxing.jiang@windriver.com>

* Add validation webhooks for PTP configuration

This change ensures that the format for the parameters in
ptpinterfaces and ptpinstances is valid when applying the
deployment configuration file.

Tests Passed:
 - Apply deployment config with correct syntax and formatting.
 - Apply deployment config with duplicate parameter keys in
   ptpinstance and ptpinterface. (Received Error)
 - Apply deployment config with no parameter key. (Received Error)
 - Apply deployment config where ptp parameters are missing "=".
   (Received Error)

Signed-off-by: Melissa Wang <melissa.wang@windriver.com>

* Accept multiple PTPInterface association

Previously, we only accept one PTP interface associated with an
interface. This commit enables the interface can accept multiple PTP
interface association.

Test:
1 Modified the deployment configuration, assigned two PTP interfaces
to an interface.
2 Bootstrap a AIOSX, deploy this node with the DM and modified
configuration
3 After the deployment, the AIOSX can be correctly deployed and
unlocked.
4 Check the PTP interfaces, both of them are assigned to a single
interface.

Signed-off-by: Yuxing Jiang <yuxing.jiang@windriver.com>

* Update deploy tool to populate ptp per instance config

This commit updates the deploy tool to correctly populate the ptp
instances, ptp interfaces and the host profile with those ptp per
instance configurations.

Test:
Deployed an AIOSX system with the PTP instances, interfaces and
parameters configured. Use the new deploy binary to generate the
deployment configurations with "./deploy build -n deployment -s vbox
--minimal-config". The new deployment-config.yaml contains those
PTP-related resources correctly configured.

Signed-off-by: Yuxing Jiang <yuxing.jiang@windriver.com>

* Update PTP Instance unit test case

This change adds ptp instance parameters to the automated tests

Signed-off-by: Melissa Wang <melissa.wang@windriver.com>

* Update version to 2.0.8-1 for charts tarball

Signed-off-by: Melissa Wang <melissa.wang@windriver.com>

Co-authored-by: Yuxing Jiang <43456284+yjian118@users.noreply.github.com>
Co-authored-by: Yuxing Jiang <yuxing.jiang@windriver.com>