Bug fix: fix error in NewPTPInterface
This commit is a bug fix to fix an error in the NewPTPInterface func,
the kind of a new PTP Interface should be the PtpInterface rather than
the PtpInstance.

Tested by generate the deployment configuration with the new DM image.
The kind of PTP interface is corrected in the new configuration.

Signed-off-by: Yuxing Jiang <yuxing.jiang@windriver.com>