Support host configurable max_cpu_frequency
This commit adds maxCPUFrequency in the hostprofile resource to support
the configuration of the the max_cpu_frequency on a host node.

Tests Passed:
1 Deploy a bare metal AIOSX w/o maxCPUFrequency configured. Generate
the deployment config with the deploy tool, the maxCPUFrequency is not
recorded.
2 Deploy a bare metal AIOSX with maxCPUFrequency configured. Validate
the max_cpu_frequncey value in the host post deploy. Generate the
deployment config with the deploy tool, the maxCPUFrequency is
recorded as configured.

Signed-off-by: Yuxing Jiang <yuxing.jiang@windriver.com>