Support validating internal list items on list types
When using kubernetes-sigs/controller-tools v2 and above, there's a
know issue that the validation on intertal list items is no longer
supported.
See in: https://github.com/kubernetes-sigs/controller-tools/issues/342.
This issue is opened 2 years ago and currently frozen(seems will not
fix in a short term). We can set a alia of the item type to introduce
the validation instead, otherwise, we'll need to remove these validations.

This commit gives an example about the effort to set the validation with
alia type on the subfunctions for review. Note: there's still many
validation of the items in a "StringList" need to implement.

Signed-off-by: Yuxing Jiang <Yuxing.Jiang@windriver.com>