Update DM playbook to accept random name of pod
When we build DM with kubebuilder v3, the manager is deployed as
deployment compared with as statefulsets in kubebuilder v1. As a
result the name of the manager pod is attached with a random id. This
commit modifies the DM playbook to search for the pod name in the
namespace, and use that name to replace hard-coded pod name in the
playbook.

Tests passed:
1. Deploy the central cloud with this playbook
2. Deploy the subcloud with this playbook by dcmanager subcloud add
3. Reconfig the subcloud with this playbook by dcmanager subcloud
reconfig

Signed-off-by: Yuxing Jiang <yuxing.jiang@windriver.com>