Migrate golangci-lint
https://install.goreleaser.com/github.com/golangci/golangci-lint.sh is not valid any more.
Need to update the DEVELOPER.md and Dockerfile.builder which refer it.
This commit opts https://raw.githubusercontent.com/golangci/golangci-lint/master/install.sh v1.13
for the stable/v2.0.8 branch. This commit also contains some change to fit for this golangci-lint.

Regarding the master branch, as we are using go v1.17, we'd better to choose a later version
of golangci-lint, and make changes to adapt the requirement of the new go version.

Test: Created new image with "make" and "make docker-build"

Signed-off-by: Yuxing Jiang <yuxing.jiang@windriver.com>