Test code migration
When we migrated DM from v1 to v3, we left test code migration
because of the limited time.
This is test code migration from the kubebuilder v1 base to v3 base.

kubebuilder v3 created the test template with ginkgo and gomega.
So that the existing test code is modified to adapt to this environment
but no new testing scenario is added in this patch

Test Plan:
PASS: Run "make test" and finished successfully
PASS: Run "make && DEBUG=yes make docker-build" and finished successfully
PASS: Confirm test coverage is similar as before migration

Signed-off-by: Takamasa Takenaka <takamasa.takenaka@windriver.com>