Fix addresses creation for any interface class (#116)
When the deployment-manager configures addresses for an
interface the DM only checks if the interface needs to
change mode to static only when it is of class data. This
commit verifies if exist static addresses bound to the interface
and change the mode to static.

Test Plan:
1. Use DM with an interface configured with class != data
2. Use DM with static addresses bound to the interface
3. Apply the DM and verify the interfaces and addresses of the host
- Interface must have the ipv4_mode/ipv6_mode set to static
- Addresses created normally

Signed-off-by: Hugo Brito <hugo.brito@windriver.com>
(cherry picked from commit dd2eb65365187a17cd950f03c10d48b48165fd6c)