Update documents
Update the following documents:
- README.md
- DEBUGGING.md
- DEVELOPER.md

Update go version and remove unnecessary packages like dep.
Add a small description how to include vender private fork.
Directory structure for example stays as in legacy structure
but mentions it is not limited under GOPATH/src.

And kustomize is installed automatically during the build process
but kustomize at "make examples" still use the external kustomize
command. It is changed to use downloaded kustomize but "null"
is evaluated as null and it is not allowed with newer kustomize.
It is changed to "".

Test Plas:
PASS: Run "make examples" finish successfully and examples are
      generated.

Signed-off-by: Takamasa Takenaka <takamasa.takenaka@windriver.com>