Frozen go-delve/delve version to v1.2.0
A recent commit in go-delve/delve breaks the support for legacy
go version. To support go ver 1.12.9, this commit freezes the version
of go-delve/delve to v1.2.0.
Note: may need to cherry-pick this commit whenever we cherry-pick
any commits to the branches based on golang 1.12.

Test plan:
Passed: built new DM image and tested to deploy an AIOSX node.

Signed-off-by: Yuxing Jiang <Yuxing.Jiang@windriver.com>