Fix addresses creation for any interface class (#136)
* Migrate golangci-lint

https://install.goreleaser.com/github.com/golangci/golangci-lint.sh is not valid any more.
Need to update the DEVELOPER.md and Dockerfile.builder which refer it.
This commit opts https://raw.githubusercontent.com/golangci/golangci-lint/master/install.sh v1.13
for the stable/v2.0.8 branch. This commit also contains some change to fit for this golangci-lint.

Regarding the master branch, as we are using go v1.17, we'd better to choose a later version
of golangci-lint, and make changes to adapt the requirement of the new go version.

Test: Created new image with "make" and "make docker-build"

Signed-off-by: Yuxing Jiang <yuxing.jiang@windriver.com>

* Fix addresses creation for any interface class (#116)

When the deployment-manager configures addresses for an
interface the DM only checks if the interface needs to
change mode to static only when it is of class data. This
commit verifies if exist static addresses bound to the interface
and change the mode to static.

Test Plan:
1. Use DM with an interface configured with class != data
2. Use DM with static addresses bound to the interface
3. Apply the DM and verify the interfaces and addresses of the host
- Interface must have the ipv4_mode/ipv6_mode set to static
- Addresses created normally

Signed-off-by: Hugo Brito <hugo.brito@windriver.com>
(cherry picked from commit dd2eb65365187a17cd950f03c10d48b48165fd6c)
(cherry picked from commit be248d6cc2d3013baad234c6fca2c81902f6d938)

* Frozen go-delve/delve version to v1.2.0

A recent commit in go-delve/delve breaks the support for legacy
go version. To support go ver 1.12.9, this commit freezes the version
of go-delve/delve to v1.2.0.
Note: may need to cherry-pick this commit whenever we cherry-pick
any commits to the branches based on golang 1.12.

Test plan:
Passed: built new DM image and tested to deploy an AIOSX node.

Signed-off-by: Yuxing Jiang <Yuxing.Jiang@windriver.com>

Signed-off-by: Yuxing Jiang <yuxing.jiang@windriver.com>
Signed-off-by: Yuxing Jiang <Yuxing.Jiang@windriver.com>
Co-authored-by: Hugo Nicodemos <hugonicodemos@users.noreply.github.com>