Add temporary support for provisioning optional filesystems (#147)
Deployment manager currently does not support adding optional
filesystems as part of the deployment model. As only filesystem resizes
are supported for select filesystems, add a exception for the
'instances' and 'image-conversion' filesystems. If either does not
exist, still pass as resize request to the system inventory manager as
the inventory manager with handle a creation and a resize operation for
these filesystems.

Test Plan:
PASS - add optional filesystems to the deployment config and observe
       their creation after the host goes INSYNC=true

Signed-off-by: Robert Church <robert.church@windriver.com>

Co-Authored-By: Yuxing Jiang <yuxing.jiang@windriver.com>