Remove resource request in manifest
The current helm chart requests cpu and memeory resource for
kube-rbac-proxy, which is not supposed to.

This commit removes the resource request in the helm chart. This
commit also removes the workaround to force generate helm chart.
This commit will not generate new DM image.

Test passed:
Deploy an AIOSX system with the new helm chart.

Signed-off-by: Yuxing Jiang <Yuxing.Jiang@windriver.com>